Nuova versione Controller

+++ Desktop/JavaEs Course/PizzeriaApi/src/main/java/com/example/demo/Controller/pizzeriaController.java
@@ -5,7 +5,6 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -15,6 +14,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.demo.Model.pizzeria;
@@ -22,64 +22,74 @@ import com.example.demo.Repository.pizzeriaRepository;
 
 @RestController
 @CrossOrigin
-@RequestMapping("/pizze")
+@RequestMapping("/")
 public class pizzeriaController {
 
 	@Autowired
 	pizzeriaRepository repositoryPizza;
-	
-	
-	
-	
+
 	@GetMapping
-	public List <pizzeria> index(){
-		
-		
-		return repositoryPizza.findAll();
+	public ResponseEntity<List<pizzeria>> index(@RequestParam(name = "filtro", required = false) String filtro) {
+		List<pizzeria> result;
+
+		if (filtro != null && !filtro.isEmpty()) {
+			result = repositoryPizza.findByNomeLike("%" + filtro + "%");
+		} else {
+			result = repositoryPizza.findAll();
+		}
+
+		if (result.size() > 0) {
+			return new ResponseEntity<List<pizzeria>>(result, HttpStatus.OK);
+		} else {
+			return new ResponseEntity<List<pizzeria>>(HttpStatus.NO_CONTENT);
+		}
 	}
-	
-	
-	@GetMapping("{id}")
-	public ResponseEntity <pizzeria> detail(@PathVariable ("id") Integer id) {
-		
-		Optional<pizzeria> res=repositoryPizza.findById(id);
+
+	@GetMapping("pizza/{id}")
+	public ResponseEntity<pizzeria> detail(@PathVariable("id") Integer id) {
+
+		Optional<pizzeria> res = repositoryPizza.findById(id);
 		if (res.isPresent()) {
 			return new ResponseEntity<pizzeria>(res.get(), HttpStatus.OK);
 		} else {
-			return new ResponseEntity <pizzeria> (HttpStatus.NOT_FOUND);
+			return new ResponseEntity<pizzeria>(HttpStatus.NOT_FOUND);
 		}
-		 
-	}
-	
-	@PostMapping()
-	public pizzeria  create (@RequestBody pizzeria pizza) {
 
-		return repositoryPizza.save(pizza);
 	}
-	
-	
-	
-	@PutMapping("{id}") 
-	public pizzeria  PaggCreate (@RequestBody pizzeria pizza, @PathVariable("id") Integer id) {
 
-		pizzeria p= repositoryPizza.getReferenceById(id);
-		
-		p.setNome(p.getNome());
-		return repositoryPizza.save(p);
+	@PostMapping("pizza/create")
+	public ResponseEntity<pizzeria> create(@RequestBody pizzeria pizza) {
+
+		pizzeria createdPizza = repositoryPizza.save(pizza);
+		if (createdPizza != null) {
+			return new ResponseEntity<pizzeria>(createdPizza, HttpStatus.OK);
+		} else {
+			return new ResponseEntity<pizzeria>(HttpStatus.BAD_REQUEST);
+		}
 	}
-	
-	
-	
-	
-	@DeleteMapping("{id}") 
-	public void delete (@PathVariable ("id") Integer id) {
-		
-		repositoryPizza.deleteById(id);
+
+	@PutMapping("pizza/{id}")
+	public ResponseEntity<pizzeria> paggCreate(@RequestBody pizzeria pizza, @PathVariable("id") Integer id) {
+
+		Optional<pizzeria> result = repositoryPizza.findById(id);
+
+		if (!result.isPresent() || result.get().getId() != pizza.getId()) {
+			return new ResponseEntity<pizzeria>(HttpStatus.BAD_REQUEST);
+		} else {
+			repositoryPizza.save(pizza);
+			return new ResponseEntity<pizzeria>(pizza, HttpStatus.OK);
+		}
+	}
+
+	@DeleteMapping("{id}")
+	public ResponseEntity<Void> delete(@PathVariable("id") Integer id) {
+
+		Optional<pizzeria> pizzaResult = repositoryPizza.findById(id);
+		if (pizzaResult.isPresent()) {
+			repositoryPizza.deleteById(id);
+			return new ResponseEntity<Void>(HttpStatus.NO_CONTENT);
+		} else {
+			return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
+		}
 	}
 }
-    
-    
-
-
-
-
